+++ Projects/Share/AD_Utils/Project.swift
@@ -2,12 +2,8 @@ import ProjectDescription
 import ProjectEnvironment
 
 let project = Project.makeModule(
-  module: .AD_OnBoarding,
+  module: .AD_Utils,
   platform: .iOS,
   product: .staticFramework,
-  dependencies: [
-    .AD_Utils,
-    .ComposableArchitecture
-  ],
   withTest: false
 )
